requirements.txt
python-dotenv>=1.0.1
langchain>=0.2.6
langchain-openai>=0.1.13
langchain-community>=0.2.6
pinecone-client>=4.1.1
pytest>=8.2.2
faiss-cpu>=1.8.0.post1